feat: Add stats plot command
