+++ pyproject.toml
@@ -23,6 +23,9 @@ sh = ">=2.0.0"
 pyyaml = ">=6"
 colorama = ">=0.4.6"
 python-dotenv = ">=0.21.1"
+textual = ">=0.48"
+rich = ">=13.7"
+plotille = ">=5"
 
 [tool.poetry.group.dev.dependencies]
 pytest = ">=7.3.2"
