+++ src/sugar/logs.py
@@ -17,6 +17,7 @@ class KxgrErrorType(Enum):
     KXGR_MISSING_PARAMETER = 6
     KXGR_INVALID_CONFIGURATION = 7
     KXGR_ACTION_NOT_IMPLEMENTED = 8
+    KXGR_NO_SERVICES_RUNNING = 9
 
 
 class KxgrLogs:
