+++ src/sugar/plugins/__init__.py
@@ -1,0 +1,1 @@
+"""Plugins module."""
